+++ options/defaults.lua
@@ -297,6 +297,9 @@ SAO.defaults = {
                 [60513] = { -- Healing Trance / Soul Preserver (Wrath)
                     [0] = false,
                 },
+                [88819] = { -- Daybreak (Cataclysm)
+                    [0] = true,
+                },
             },
             glow = {
                 [879] = { -- Exorcism
@@ -323,6 +326,9 @@ SAO.defaults = {
                 [59578] = { -- The Art of War (2/2) (Wrath+)
                     [879]   = true, -- Exorcism
                     [19750] = true, -- Flash of Light (not for Cata)
+                },
+                [88819] = { -- Daybreak (Cataclysm)
+                    [20473] = true, -- Holy Shock
                 },
             },
         },